refactor(oauth): P3 - decompose oauth-install into focused internal modules
Split the 700-line oauth-install.ts monolith into:
- oauth-install-bundle.ts: JAR deploy and OSGi config writing
- oauth-install-gogo.ts: Gogo shell command execution and key-value parsing
- oauth-install-verify.ts: post-provisioning token verification

oauth-install.ts is now a thin facade with public types and orchestration.
